+++ concreate/descriptor.py
@@ -1,11 +1,10 @@
-import hashlib
 import logging
 import shutil
-import subprocess
 import os
 
 from concreate import DEFAULT_USER, tools
 from concreate.errors import ConcreateError
+from concreate.resource import Resource
 from concreate.version import schema_version
 
 logger = logging.getLogger('concreate')
@@ -24,6 +23,8 @@ class Descriptor(object):
                                                 descriptor_type)
         if descriptor_type == 'image':
             self.check_schema_version()
+
+        self.module_repositories = {}
 
     def check_schema_version(self):
         """ Check supported schema version """
@@ -47,6 +48,9 @@ class Descriptor(object):
     def items(self):
         return self.descriptor.items()
 
+    def get(self, k, default=None):
+        return self.descriptor.get(k, default)
+
     def label(self, key):
         for l in self.descriptor['labels']:
             if l['name'] == key:
@@ -61,8 +65,7 @@ class Descriptor(object):
             self._process_execute()
         if 'ports' in self.descriptor:
             self._process_ports()
-        if 'dependencies' in self.descriptor:
-            self._process_dependencies()
+        self._process_module_repositories()
         self._process_run()
         self._process_labels()
         return self
@@ -116,13 +119,10 @@ class Descriptor(object):
         for port in self.descriptor['ports']:
             port['name'] = port['value']
 
-    def _process_dependencies(self):
-        """ Generate name attribute for dependencies """
-        dependencies = {}
-        for dependency in self.descriptor['dependencies']:
-            resource = Resource.new(dependency, self.directory)
-            dependencies[resource.name] = resource
-        self.descriptor['dependencies'] = dependencies
+    def _process_module_repositories(self):
+        for repo in self.descriptor.get('modules', {}).get('repositories', []):
+            resource = Resource.new(repo, self.directory)
+            self.module_repositories[resource.name] = resource
 
     def _process_labels(self):
         """ Generate labels from concreate keys """
@@ -149,163 +149,5 @@ class Descriptor(object):
         if not self.label('summary') and description:
             self.descriptor['labels'].append(
                 {'name': 'summary', 'value': description['value']})
-
-
-class Resource(object):
-    SUPPORTED_HASH_ALGORITHMS = ['sha256', 'sha1', 'md5']
-    CHECK_INTEGRITY = True
-
-    @staticmethod
-    def new(resource, base_dir=os.getcwd()):
-        if 'git' in resource:
-            return GitResource(resource)
-        elif 'path' in resource:
-            directory = resource['path']
-            if not os.path.isabs(directory):
-                resource['path'] = os.path.join(base_dir, directory)
-            return PathResource(resource)
-        elif 'url' in resource:
-            return UrlResource(resource)
-        raise ValueError("Resource type is not supported: %s" (resource))
-
-    def __init__(self, descriptor):
-        self.name = descriptor['name']
-        if 'description' in descriptor:
-            self.desription = descriptor['description']
-        self.checksums = {}
-        for algorithm in Resource.SUPPORTED_HASH_ALGORITHMS:
-            if algorithm in descriptor:
-                self.checksums[algorithm] = descriptor[algorithm]
-
-    def _copy_impl(self, target):
-        raise NotImplementedError("Implement _copy_impl() for Resource: "
-                                  + self.__module__ + "."
-                                  + type(self).__name__)
-
-    def target_file_name(self):
-        return os.path.basename(self.name)
-
-    def copy(self, target_dir=os.getcwd()):
-        target = os.path.join(target_dir, self.target_file_name())
-
-        overwrite = False
-        if os.path.exists(target):
-            try:
-                self.__verify(target)
-            except:
-                overwrite = True
-
-        if os.path.exists(target) and not overwrite:
-            logger.debug("Resource '%s' exists, skipping" % self.name)
-            return target
-
-        try:
-            self._copy_impl(target)
-        except Exception as ex:
-            # exception is fatal we be logged before Concreate dies
-            raise ConcreateError('Error copying Resource: %s' % (self.name),
-                                 ex)
-
-        self.__verify(target)
-
-        return target
-
-    def __verify(self, target):
-        """ Checks all defined check_sums for an aritfact """
-        if not self.checksums:
-            return True
-        if not Resource.CHECK_INTEGRITY:
-            logger.info("Integrity checking disabled, skipping verification.")
-            return True
-        if os.path.isdir(target):
-            logger.info("Target is directory, cannot verify checksum.")
-            return True
-        for algorithm, checksum in self.checksums.items():
-            self.__check_sum(target, algorithm, checksum)
-        return True
-
-    def __check_sum(self, target, algorithm, expected):
-        """ Check that file chksum is correct
-        Args:
-          alg - algorithm which will be used for diget
-          expected_chksum - checksum which artifact must mathc
-        """
-
-        logger.debug("Checking '%s' %s hash..." % (self.name, algorithm))
-
-        hash_function = getattr(hashlib, algorithm)()
-
-        with open(target, "rb") as f:
-            for chunk in iter(lambda: f.read(65536), b""):
-                hash_function.update(chunk)
-        checksum = hash_function.hexdigest()
-
-        if checksum.lower() != expected.lower():
-            raise ConcreateError("The %s computed for the '%s' file ('%s') doesn't match the '%s' value"  # noqa: E501
-                                 % (algorithm, self.name, checksum, expected))
-
-        logger.debug("Hash is correct.")
-
-
-class PathResource(Resource):
-    def __init__(self, descriptor):
-        if 'name' not in descriptor:
-            descriptor['name'] = os.path.basename(descriptor['path'])
-        super(PathResource, self).__init__(descriptor)
-        self.path = descriptor['path']
-
-    def _copy_impl(self, target):
-        if (os.path.isdir(self.path)):
-            shutil.copytree(self.path, target)
-        else:
-            shutil.copy(self.path, target)
-        return target
-
-
-class UrlResource(Resource):
-    def __init__(self, descriptor):
-        if 'name' not in descriptor:
-            descriptor['name'] = os.path.basename(descriptor['url'])
-        super(UrlResource, self).__init__(descriptor)
-        self.url = descriptor['url']
-
-    def _copy_impl(self, target):
-        tools.download_file(self.__substitute_cache_url(), target)
-        return target
-
-    def __substitute_cache_url(self):
-        # XXX: maybe put this in [common]
-        cache = tools.cfg.get('artifact', {}).get('cache_url', None)
-        if not cache:
-            return self.url
-
-        for algorithm in Resource.SUPPORTED_HASH_ALGORITHMS:
-            if algorithm in self.checksums:
-                logger.debug("Using %s to fetch artifacts from cacher."
-                             % algorithm)
-                return (cache.replace('#filename#', self.name)
-                        .replace('#algorithm#', algorithm)
-                        .replace('#hash#', self.checksums[algorithm]))
-        return self.url
-
-
-class GitResource(Resource):
-    def __init__(self, descriptor):
-        if 'name' not in descriptor:
-            descriptor['name'] = os.path.basename(descriptor['git']['url'])
-        super(GitResource, self).__init__(descriptor)
-        self.url = descriptor['git']['url']
-        self.ref = descriptor['git']['ref']
-
-    def target_file_name(self):
-        # XXX: We could make a case for using name instead of repo-ref
-        return "%s-%s" % (os.path.basename(self.url), self.ref)
-
-    def _copy_impl(self, target):
-        cmd = ['git', 'clone', '--depth', '1', self.url, target, '-b',
-               self.ref]
-        logger.debug("Running '%s'" % ' '.join(cmd))
-        subprocess.check_output(cmd, stderr=subprocess.STDOUT)
-        return target
